refactor(native): promote lance_native_planner to module dir; add NativeDirection
